Add config option for equivalent strings
