+++ answerset/config.json
@@ -2,6 +2,7 @@
     "Enable Answer Choice Comments [...]": false,
     "Enable Answer Comments (...)": false,
     "Enable Lenient Validation": true,
+    "Equivalent Strings": [],
     "Ignore Case": true,
     "Ignore Separators in Brackets": true,
     "Ignored Characters": " .-"
