+++ answerset/group.py
@@ -1,0 +1,28 @@
+import unicodedata as ucd
+
+def is_combining(ch: str) -> bool:
+    """Check if a character is a Unicode combining mark."""
+
+    return ucd.category(ch).startswith('M')
+
+def group_combining(string: str) -> list[str]:
+    """Group combining characters with the previous character."""
+
+    parts = []
+    current = ''
+    for ch in string:
+        if is_combining(ch):
+            if not current:
+                current = '\xa0'
+
+            current += ch
+        else:
+            if current:
+                parts.append(current)
+
+            current = ch
+
+    if current:
+        parts.append(current)
+
+    return parts
